added sample auth using api-key for user controller using auth function using developer service

+++ src/main/java/plbtw/klmpk/barang/hilang/controller/UserController.java
@@ -4,6 +4,7 @@
  */
 package plbtw.klmpk.barang.hilang.controller;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,12 +13,15 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import plbtw.klmpk.barang.hilang.entity.User;
+import plbtw.klmpk.barang.hilang.entity.form.request.ApiKeyRequest;
 import plbtw.klmpk.barang.hilang.entity.form.request.UserRequest;
 import plbtw.klmpk.barang.hilang.message.CustomResponseMessage;
+import plbtw.klmpk.barang.hilang.service.DeveloperService;
 import plbtw.klmpk.barang.hilang.service.RoleService;
 import plbtw.klmpk.barang.hilang.service.UserService;
 
@@ -30,29 +34,72 @@ import plbtw.klmpk.barang.hilang.service.UserService;
 public class UserController {
   @Autowired
   UserService userService;
+  @Autowired
+  DeveloperService developerService;
 
+  private boolean authApiKey(String token) {
+    if (developerService.getDeveloperByApiKey(token) == null)
+      return false;
+    return true;
+  }
 
   @RequestMapping(method = RequestMethod.GET, produces = "application/json")
-  public Collection<User> getAllUsers() {
-    List<User> allUsers = (List<User>) userService.getAllUsers();
-    for (User user : allUsers) {
-      Link selfLink = linkTo(UserController.class).withSelfRel();
-      user.add(selfLink);
+  public CustomResponseMessage getAllUsers(@RequestHeader String apiKey) {
+    try {
+      if (!authApiKey(apiKey)) {
+        return new CustomResponseMessage(HttpStatus.FORBIDDEN,
+            "Please use your api key to authentication");
+      }
+
+      List<User> allUsers = (List<User>) userService.getAllUsers();
+      for (User user : allUsers) {
+        Link selfLink = linkTo(UserController.class).withSelfRel();
+        user.add(selfLink);
+      }
+      CustomResponseMessage result = new CustomResponseMessage();
+      result.add(linkTo(UserController.class).withSelfRel());
+      result.setHttpStatus(HttpStatus.ACCEPTED);
+      result.setMessage("Success");
+      result.setResult(allUsers);
+      return result;
+    } catch (Exception ex) {
+      return new CustomResponseMessage(HttpStatus.FORBIDDEN,
+          "Please use your api key to authentication");
     }
-    return allUsers;
   }
 
   @RequestMapping(value = "/find/{id}", method = RequestMethod.GET, produces = "application/json")
-  public User getUser(@PathVariable("id") long id) {
-    User user = userService.getUser(id);
-    Link selfLink = linkTo(UserController.class).withSelfRel();
-    user.add(selfLink);
-    return user;
+  public CustomResponseMessage getUser(@RequestHeader String apiKey, @PathVariable("id") long id) {
+    try {
+      if (!authApiKey(apiKey)) {
+        return new CustomResponseMessage(HttpStatus.FORBIDDEN,
+            "Please use your api key to authentication");
+      }
+      User user = userService.getUser(id);
+      Link selfLink = linkTo(UserController.class).withSelfRel();
+      user.add(selfLink);
+      List<User> listUser = new ArrayList<User>();
+      listUser.add(user);
+      CustomResponseMessage result = new CustomResponseMessage();
+      result.add(linkTo(UserController.class).withSelfRel());
+      result.setHttpStatus(HttpStatus.ACCEPTED);
+      result.setMessage("Success");
+      result.setResult(listUser);
+      return result;
+    } catch (Exception ex) {
+      return new CustomResponseMessage(HttpStatus.FORBIDDEN,
+          "Please use your api key to authentication");
+    }
   }
 
   @RequestMapping(method = RequestMethod.POST, produces = "application/json")
-  public CustomResponseMessage addUser(@RequestBody UserRequest userRequest) {
+  public CustomResponseMessage addUser(@RequestHeader String apiKey,
+      @RequestBody UserRequest userRequest) {
     try {
+      if (!authApiKey(apiKey)) {
+        return new CustomResponseMessage(HttpStatus.FORBIDDEN,
+            "Please use your api key to authentication");
+      }
       User user = new User();
       user.setUsername(userRequest.getUsername());
       user.setEmail(userRequest.getEmail());
@@ -67,8 +114,13 @@ public class UserController {
   }
 
   @RequestMapping(method = RequestMethod.PUT, produces = "application/json")
-  public CustomResponseMessage updateUser(@RequestBody UserRequest userRequest) {
+  public CustomResponseMessage updateUser(@RequestHeader String apiKey,
+      @RequestBody UserRequest userRequest) {
     try {
+      if (!authApiKey(apiKey)) {
+        return new CustomResponseMessage(HttpStatus.FORBIDDEN,
+            "Please use your api key to authentication");
+      }
       User userUpdate = userService.getUser(userRequest.getId());
       userUpdate.setUsername(userRequest.getUsername());
       userUpdate.setAlamat(userRequest.getAlamat());
@@ -83,8 +135,13 @@ public class UserController {
   }
 
   @RequestMapping(method = RequestMethod.DELETE, produces = "application/json")
-  public CustomResponseMessage deleteUser(@RequestBody UserRequest userRequest) {
+  public CustomResponseMessage deleteUser(@RequestHeader String apiKey,
+      @RequestBody UserRequest userRequest) {
     try {
+      if (!authApiKey(apiKey)) {
+        return new CustomResponseMessage(HttpStatus.FORBIDDEN,
+            "Please use your api key to authentication");
+      }
       userService.deleteUser(userRequest.getId());
       return new CustomResponseMessage(HttpStatus.CREATED, "Delete Successful");
     } catch (Exception ex) {
